// types/characterTypes.tsx
export interface characterInterface {
    info: {
        count: number,
        pages: number,
        next: string | null
        prev: string | null
    }
    results: characterInfoInterface[]
}
export interface characterInfoInterface {
    id: number,
    name: string
    status: string
    species: string
    type: string
    gender: string
    origin: {
        name: string
        url: string
    }
    location: {
        name: string
        url: string
    }
    image: string
    episode: string[]
    url: string
    created: string
}